Add turnos
Falta manejar cuando todos los gusanos mueren y un par de cosas mas. Si se esta moviendo cuando termina el turno, se queda quieto en el lugar pero sigue la animacion de caminar. No se si es algo del cliente o del server

+++ src/server/game/server_gamestatus.cpp
@@ -11,8 +11,7 @@ GameStatusError::GameStatusError(uint8_t code, GameWorld& world): code(code), Ga
 void GameStatusError::serialize(ServerProtocol& protocol) { protocol.send_game_errormessage(code); }
 
 
-GameStatusScenario::GameStatusScenario(GameWorld& world):
-        GameStatus(world) {
+GameStatusScenario::GameStatusScenario(GameWorld& world): GameStatus(world) {
     height = width = 0;
     world.get_dimensions(&height, &width);
 }
@@ -21,11 +20,12 @@ void GameStatusScenario::serialize(ServerProtocol& protocol) {
     protocol.send_scenario(entities_info, height, width);
 }
 
-GameStatusStart::GameStatusStart(GameWorld& world): worms_info(world.get_worms_info()), GameStatus(world) {}
+GameStatusStart::GameStatusStart(GameWorld& world):
+        worms_info(world.get_worms_info()), GameStatus(world) {}
 
 void GameStatusStart::serialize(ServerProtocol& protocol) { protocol.send_worms_list(worms_info); }
 
-GameStatusRunning::GameStatusRunning(uint8_t current_worm, GameWorld& world):
+GameStatusRunning::GameStatusRunning(uint16_t current_worm, GameWorld& world):
         current_worm(current_worm), GameStatus(world) {}
 
 void GameStatusRunning::serialize(ServerProtocol& protocol) {
